fix: Moved code to a new src directory, fixed a bug in PDFExtractor in reducing chunk size to 1000 instead of 4000 and updated docs
