+++ src/components/PDFExtractor.py
@@ -26,7 +26,7 @@ class PDFExtractor:
                         raw_text += text
 
                 text_splitter = RecursiveCharacterTextSplitter(        
-                    chunk_size = 4000,
+                    chunk_size = 1000,
                     chunk_overlap  = 200,
                     length_function = len,
                 )
