+++ src/components/PDFExtractorEnum.py
