+++ src/main.py
@@ -7,7 +7,7 @@ from PyPDF2 import PdfReader
 from langchain.embeddings.openai import OpenAIEmbeddings
 from langchain.text_splitter import CharacterTextSplitter
 from langchain.vectorstores import FAISS
-from openai_utils import get_openai_api_key
+from openai_customdoc_search.src.openai_utils import get_openai_api_key
 from langchain.chains.question_answering import load_qa_chain
 from langchain.llms import OpenAI
